excercise

+++ NestedLoops/Trip.py
@@ -1,10 +1,10 @@
 destination = input()
 
 while destination != "End":
-    need_money = int(input())
+    need_money = float(input())
     money = 0
     while money < need_money:
-        saved_money = int(input())
+        saved_money = float(input())
         money += saved_money
     else:
         print(f"Going to {destination}!")
